Move cache:rebuild to own command file
https://drupal.slack.com/archives/C62H9CWQM/p1694091163310429?thread_ts=1693915858.778809&cid=C62H9CWQM
