+++ tests/functional/SqlDumpTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Unish;
 
 use Drush\Commands\core\CacheCommands;
+use Drush\Commands\core\CacheRebuildCommands;
 use Drush\Commands\sql\SqlCommands;
 use Symfony\Component\Filesystem\Path;
 
@@ -38,7 +39,7 @@ class SqlDumpTest extends CommandUnishTestCase
         ];
 
         // In Drupal 9.1+, cache_discovery et. al. do not exist until after a cache rebuild.
-        $this->drush(CacheCommands::REBUILD, []);
+        $this->drush(CacheRebuildCommands::REBUILD, []);
 
         $this->drush(SqlCommands::DUMP, [], $options + ['simulate' => null]);
         $expected = $this->dbDriver() == 'mysql' ? '--ignore-table=unish_dev.cache_discovery' : '--exclude-table=cache_discovery';
